+++ src/resources/hubs/hub.ts
@@ -47,12 +47,6 @@ export interface BillingProfile {
   term: Term;
   allow_prepaid?: boolean;
   disable?: boolean;
-  trial: Trial | null;
-}
-
-export interface Trial {
-  start: Time;
-  end: Time;
 }
 
 export interface HubMetas {